+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - New plugin Cart
+
+### Removed
+ - Plugin Basket
+ 
+## [2.8.1] - 2017-12-19
+### Fixed
+ - Form with paragraph throws RNG error
+
+## [2.8.0] - 2017-12-17
+### Added
  - Domain specific meta robots values support
 
 ### Changed
@@ -289,6 +300,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - System messages from previous request are now marked via class.
 
 [Unreleased]: https://bitbucket.org/igwr/cms/compare/dev..master
+[2.8.1]: https://bitbucket.org/igwr/cms/compare/v2.8.1..v2.8.0
+[2.8.0]: https://bitbucket.org/igwr/cms/compare/v2.8.0..v2.7.2
 [2.7.2]: https://bitbucket.org/igwr/cms/compare/v2.7.2..v2.7.1
 [2.7.1]: https://bitbucket.org/igwr/cms/compare/v2.7.1..v2.7.0
 [2.7.0]: https://bitbucket.org/igwr/cms/compare/v2.7.0..v2.6.2